refactor: reorganize code & minor linting fixes
