+++ functions/.eslintrc.js
@@ -1,19 +1,25 @@
 module.exports = {
+  root: true,
   env: {
     es6: true,
     node: true,
   },
+  parser: "@typescript-eslint/parser",
   parserOptions: {
-    "ecmaVersion": 2022,
+    ecmaVersion: 2022,
+    sourceType: "module",
+    project: "./tsconfig.json",
   },
+  plugins: ["@typescript-eslint"],
   extends: [
     "eslint:recommended",
-    "google",
+    "plugin:@typescript-eslint/recommended",
   ],
   rules: {
     "no-restricted-globals": ["error", "name", "length"],
     "prefer-arrow-callback": "error",
     "quotes": ["error", "double", {"allowTemplateLiterals": true}],
+    "@typescript-eslint/no-unused-vars": ["error", { "argsIgnorePattern": "^_", "varsIgnorePattern": "^_", "caughtErrorsIgnorePattern": "^_" }],
   },
   overrides: [
     {
@@ -25,4 +31,5 @@ module.exports = {
     },
   ],
   globals: {},
+  ignorePatterns: ["lib/**/*", "node_modules/**/*"],
 };
